projects page content fields added to cms
